Move connection indicators into its own MVC setup
No more God class for the main window controller!
