+++ app/src/main/resources/edu/wpi/first/shuffleboard/app/ConnectionIndicators.fxml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.HBox?>
+<HBox xmlns="http://javafx.com/javafx"
+      xmlns:fx="http://javafx.com/fxml"
+      fx:controller="edu.wpi.first.shuffleboard.app.ConnectionIndicatorsController"
+      spacing="2" alignment="CENTER"
+      styleClass="connection-indicator-area"
+      fx:id="root">
+
+</HBox>
